pyproject.toml
[project]
name = "spotinator"
version = "0.1.0"
description = ""
authors = [{ name = "manav", email = "manavjotsingh17@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
  "fastapi[all] (>=0.115.12,<0.116.0)",
  "sqlmodel (>=0.0.24,<0.0.25)",
  "langchain-mcp-adapters (>=0.0.7,<0.0.8)",
  "requests (>=2.32.3,<3.0.0)",
  "langchain-google-genai (>=2.1.2,<3.0.0)",
  "langchain-community (>=0.3.21,<0.4.0)",
  "langgraph-checkpoint-sqlite (>=2.0.6,<3.0.0)",
  "langgraph (>=0.3.25,<0.4.0)",
]

[tool.poetry]
packages = [{ include = "spotinator", from = "src" }]

[tool.poetry.scripts]
spot = "spotinator.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
